Add files via upload
Microscopic heat
engines can exploit thermal as well as active fluctuations to extract thermodynamic work. We investigate the
properties of a microscopic Stirling’s engine that uses an active (self-propelling) particle as a working substance,
in contact with two thermal baths. It is shown that the presence of activity leads to an enhanced performance of
the engine. The efficiency can be improved by increasing the activity strength for all cycle time, including the
nonquasistatic regime.We verify that the analytical results agree very well with our simulations. The variation of
efficiency with the temperature difference between the two thermal baths has also been explored. The optimum
region of operation of the engine has been deduced, by using its efficient power (product of efficiency and power)
as a quantifier. Finally, a simple model is provided that emulates the behavior of a flywheel driven by this engine.